asset-updater/requirements.txt
pyyaml==5.4.1
query-string==2019.4.13
requests==2.22.0
typepy==1.1.1
urllib3==1.25.3
pyyaml==5.4.1